Move IntPair in jflex.base
