+++ jflex/src/main/java/jflex/base/IntPair.java
@@ -7,7 +7,7 @@
  *                                                                         *
  * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * */
 
-package jflex.core;
+package jflex.base;
 
 /**
  * Simple pair of integers.
